chore: update vss (#9)
* Simplify some code

* use `Result::unwrap_or_default()`
* use `str::strip_prefix()`

* build: update jOOQ & fix Gradle 9.0 compatibility issues

With the release of Gradle 9.0.0, we experienced build
failures due to deprecated access to plugin conventions
(JavaPluginConvention).
Ref: https://docs.gradle.org/9.0.0-rc-1/userguide/upgrading_version_8.html#deprecated_access_to_conventions

To address this, we:
1. Upgrade jOOQ Gradle plugin from 8.0 to 8.2.2 to
avoid deprecation warnings when using Gradle 9.0.
Ref: https://github.com/etiennestuder/gradle-jooq-plugin/releases/tag/v8.2.2

2. Upgrade jOOQ from 3.17.7 to 3.19.1 to match the
plugin version upgrade.

Note: We avoid the latest version of jOOQ because it
requires Java 21.

* build: add JUnit platform launcher for test execution

Additionally, we add JUnit Platform Launcher which
is required for JUnit 5 tests in newer Gradle versions.

* ci: update Gradle to 9.0

update to Gradle 9.0 and auto-extract version
number.

* Fix typos in readme files

* Create and initialize the database if it does not exist

Also implement a versioning and migration scheme for future updates to
the schema. It is an adaptation of the scheme used in CLN.

* Let the test suite setup the database

* Add migrations tests

* Move `MIGRATIONS` list into dedicated `migrations` module

* Add deprecation notice of the Java-based implementation

* Mark Rust-based implementation of VSS server as ready to use

Also remove mentions of the Java-based implementation in the main
README.

* Update `Cargo.lock` for current Rust stable

.. especially given that there is a new non-forwards-compatible version.

* Use Rust version in CI

* Add LDK Node integration tests

While we already have VSS integration tests on the LDK Node side, we
here also run them here to ensure any *server* changes do not result in
test failures.

* Rename `server` crate to `vss-server`

* Use appropriate HTTP status codes in error responses

Previously, we would just always return
`StatusCode::INTERNAL_SERVER_ERROR`. Here we fix this behavior and
implment what we did in the now-deprecated Java version.

* Spell out `InternalServerError`

* Add `rustfmt` to CI to ensure formatting is checked

.. we simply add a `cargo fmt` step to our Rust CI

* Fix visibility errors in `postgres_store` tests

* Add the option to make TLS connections to the database

Also add the option to specify an additional root certificate for
the client to trust.

* Add option to verify JWT tokens in the HTTP Authorization header

The RSA public key against which the JWT tokens are verified can be set
either via a configuration file setting, or an environment variable,
with the latter having the higher priority.

---------

Co-authored-by: Andrei <andrei.i@posteo.de>
Co-authored-by: Gursharan Singh <3442979+G8XSU@users.noreply.github.com>
Co-authored-by: Enigbe <ochekliyeenigbe@gmail.com>
Co-authored-by: Jeffrey Czyz <jkczyz@gmail.com>
Co-authored-by: Elias Rohrer <dev@tnull.de>
Co-authored-by: Leo Nash <hello@leonash.net>